Created Status Bar for Build Targets :
Status bar now shows the build targets much like the SIMPL+ IDE, It will change the build types depending on the file (build types are tracked on the USH file when the UDP gets compiled)

Introduced F12 keyboard shortcut that will compile based on the build types shown in the Status Bar

Clicking on the status bar will open a quick pick window where the Build Types for that document can be changed. Build types will persist while VS Code is open, however they will revert to default or USH if the UDP is not compiled. Just like with the SIMPL+ IDE.

See action UML for functionality.